requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.5
psycopg2-binary>=2.9.9
python-decouple>=3.8
gunicorn>=21.2.0
django-filter>=23.5



numpy>=1.26
pandas>=2.2
matplotlib>=3.8
scikit-learn>=1.4
tensorflow==2.19.0
keras>=3.5.0

python-telegram-bot==22.1

Pillow>=10.3.0
whitenoise>=6.6.0

dj-database-url>=1.0.0
